% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{calcprob}
\alias{calcprob}
\title{Function to calculate vector of max entropy noise distribution for a given value of b and d}
\usage{
calcprob(b, d, n, tolentr = 1e-06, tolq = 1e-06, maxit = 20)
}
\arguments{
\item{b}{}

\item{d}{}

\item{n}{}

\item{tolentr}{}

\item{tolq}{}

\item{maxit}{}
}
\value{

}
\description{
Function to calculate vector of max entropy noise distribution for a given value of b and d
}
\author{
Johan Heldal
}
