% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellKeyViaDummy.R
\name{CellKeyViaDummy}
\alias{CellKeyViaDummy}
\title{Converting tabular data by a perturbation table and uniformly distributed keys}
\usage{
CellKeyViaDummy(
  data,
  freqVar = NULL,
  rKeyVar = NULL,
  hierarchies = NULL,
  formula = NULL,
  dimVar = NULL,
  preAggregate = is.null(freqVar),
  total = "Total",
  x = NULL,
  crossTable = NULL,
  D = 5,
  V = 3,
  js = 2,
  pstay = NULL
)
}
\arguments{
\item{data}{Input data as a data frame (inner cells)}

\item{freqVar}{Variable holding counts (name or number)}

\item{rKeyVar}{Variable holding uniformly distributed keys}

\item{hierarchies}{List of hierarchies, which can be converted by \code{\link{AutoHierarchies}}.
Thus, a single string as hierarchy input is assumed to be a total code.
Exceptions are \code{"rowFactor"} or \code{""}, which correspond to only using the categories in the data.}

\item{formula}{Model formula defining publishable cells. Will be used to calculate \code{x} (via \code{\link{ModelMatrix}}).}

\item{dimVar}{The main dimensional variables and additional aggregating variables. This parameter can be  useful when hierarchies and formula are unspecified.}

\item{preAggregate}{When \code{TRUE}, the data will be aggregated within the function to an appropriate level.}

\item{total}{String used to name totals}

\item{x}{Dummy matrix defining cells to be published (possible as input instead of generated)}

\item{crossTable}{Data frame to accompany \code{x} when \code{x} is input.}

\item{D}{\code{\link{pt_create_pParams}} parameter}

\item{V}{\code{\link{pt_create_pParams}} parameter}

\item{js}{\code{\link{pt_create_pParams}} parameter}

\item{pstay}{\code{\link{pt_create_pParams}} parameter}
}
\value{
Data frame with keys or aggregated counts (original and perturbated)
}
\description{
The perturbation table is created by package ptable.
}
\details{
With \code{freqVar=NULL} and \code{preAggregate = FALSE}, the calculated keys are returned.
}
\examples{
z <- data.frame(geo  = c("Iceland", "Portugal", "Spain"), 
                eu = c("nonEU", "EU", "EU"),
                year = rep(c("2018","2019"), each = 3),
                freq = c(2,3,7,1,5,6), stringsAsFactors = FALSE)

CellKeyViaDummy(z, "freq", formula = ~eu * year + geo)
CellKeyViaDummy(z, freqVar = NULL, formula = ~eu * year + geo)
CellKeyViaDummy(z, freqVar = NULL, formula = ~eu * year + geo, preAggregate = FALSE)
z$keys <- sin(1:6)\%\%1
CellKeyViaDummy(z, "freq", "keys", formula = ~eu * year + geo)
CellKeyViaDummy(z, "freq", "keys", dimVar = c("geo", "eu", "year"))
CellKeyViaDummy(z, freqVar = NULL, "keys", dimVar = c("geo", "eu", "year"))
CellKeyViaDummy(z, "freq", "keys", hierarchies = 
     list(geo = c("EU", "@Portugal", "@Spain", "Iceland"), year = c("2018", "2019")))

# my_km2 is temporary function, SSBtoolsData('my_km2') available in next SSBtools version
my_km2 <- SSBcellKey:::my_km2()
set.seed(123)
my_km2$keys <- runif(nrow(my_km2))
CellKeyViaDummy(my_km2, "freq", "keys", formula = ~(Sex + Age) * Municipality * Square1000m + Square250m)      
}
