% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{freq_maxentropy}
\alias{freq_maxentropy}
\title{Maximum entropy noise function for frequency table perturbation using the
cell-key method}
\usage{
freq_maxentropy(noisefactor, ddc, minpositive, ...)
}
\arguments{
\item{noisefactor}{}

\item{ddc}{}

\item{minpositive}{}

\item{...}{}
}
\value{

}
\description{
Maximum entropy noise function for frequency table perturbation using the
cell-key method
}
