% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_num_lookup.R
\name{generate_lookup_table}
\alias{generate_lookup_table}
\title{Generate a lookup table given noise probabilities.}
\usage{
generate_lookup_table(probmatrix, ncellkeys, datacomponents, seed = 123, ...)
}
\arguments{
\item{probmatrix}{matrix containing noise probabilities}

\item{ncellkeys}{number of cellkeys.}

\item{datacomponents}{data-driven components used for lookup}

\item{seed}{set seed for reproducibility of random results.}
}
\value{
returns a mxn matrix, for m possible cell keys and n possible data
driven components used for lookup in the cell key method.
}
\description{
Generate a lookup table given noise probabilities.
}
\examples{
dcomponent <- seq(from = 50, to = 100, by = 10)
a4 <-  generate_prob_matrix(
    D = 15,
    step = 1,
    dcomponent,
    split_triangular,
    width = 3,
    ddc2noise = function(x)
      x / 10,
    percnoise = TRUE
  )
a <- generate_lookup_table(a4, 16, dcomponent/100)
}
\author{
Daniel P. Lupp, Hege Marie Bøvelstad
}
