% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{laplace_noise}
\alias{laplace_noise}
\title{Title}
\usage{
laplace_noise(noisefactor, ddc, split = 0, scale = 2, ...)
}
\arguments{
\item{noisefactor}{}

\item{ddc}{}

\item{split}{}

\item{scale}{}

\item{...}{}
}
\value{

}
\description{
Title
}
