% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{normal_noise}
\alias{normal_noise}
\title{Normal noise function for perturbation using the cell-key method}
\usage{
normal_noise(noisefactor, ddc, step = step, ...)
}
\arguments{
\item{noisefactor}{}

\item{ddc}{}

\item{step}{}

\item{...}{}
}
\value{

}
\description{
Normal noise function for perturbation using the cell-key method
}
