% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pTable.R
\name{Pconvert}
\alias{Pconvert}
\title{Convert counts using pTable}
\usage{
Pconvert(x, pMatrix, rkeys = runif(NROW(x)))
}
\arguments{
\item{x}{counts}

\item{pMatrix}{Output from \code{\link{Pmatrix}}}

\item{rkeys}{uniformly distributed keys}
}
\value{
converted counts
}
\description{
Convert counts using pTable
}
\examples{
Pconvert(rep(1, 4), Pmatrix(), c(0.5, 0.8, 0.9, 0.99))
Pconvert(1:9, Pmatrix(), (1:9)/10)
Pconvert(1:9, Pmatrix(), rev((1:9)/10))
}
\author{
Øyvind Langsrud
}
