% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{split_triangular}
\alias{split_triangular}
\title{Symmetric two peak triangular noise.}
\usage{
split_triangular(noisefactor, ddc, width = ddc/3, ddc2noise = NULL, ...)
}
\arguments{
\item{noisefactor}{vector of possible noise factors.}

\item{ddc}{data driven component used for lookup.}

\item{width}{numeric scalar, width of noise triangle.}

\item{ddc2noise}{function to translate \code{ddc} to the distance between
triangles and 0. If set to NULL (default), \code{ddc} is equal to distance between
triangles and 0.}

\item{...}{}
}
\value{
numeric vector describing noise distribution.
}
\description{
Symmetric two peak triangular noise.
}
\examples{
nf <- seq(from = -15, to = 15, by = 1)
p1 <- split_triangular(noisefactor = nf,
                       ddc = 5,
                       width = 4)
p2 <- split_triangular(noisefactor = nf,
                       ddc = 5,
                       ddc2noise = function(x) 2*x,
                       width = 4)
plot(x = nf, y = p1)
plot(x = nf, y = p2)
}
