% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{uniform_noise}
\alias{uniform_noise}
\title{Uniform noise function for perturbation using the cell-key method}
\usage{
uniform_noise(noisefactor, ...)
}
\arguments{
\item{noisefactor}{}

\item{...}{}
}
\value{

}
\description{
Uniform noise function for perturbation using the cell-key method
}
