% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellKeyFits.R
\name{CellKeyFits}
\alias{CellKeyFits}
\title{CellKeyFits}
\usage{
CellKeyFits(
  data,
  freqVar = NULL,
  rKeyVar = NULL,
  hierarchies = NULL,
  formula = NULL,
  dimVar = NULL,
  preAggregate = is.null(freqVar),
  extend0 = TRUE,
  iter = 1000,
  eps = 0.01,
  tol = 1e-13,
  reduceBy0 = TRUE,
  reduceByColSums = TRUE,
  reduceByLeverage = FALSE,
  ...
)
}
\arguments{
\item{data}{data frame (inner cells)}

\item{freqVar}{Variable holding counts}

\item{rKeyVar}{Variable holding keys}

\item{hierarchies}{List of hierarchies}

\item{formula}{Model formula}

\item{dimVar}{Dimensional variables}

\item{preAggregate}{Aggregation}

\item{extend0}{Extend0}

\item{iter}{Mipf parameter}

\item{eps}{Mipf parameter}

\item{tol}{Mipf parameter}

\item{reduceBy0}{Mipf parameter}

\item{reduceByColSums}{Mipf parameter}

\item{reduceByLeverage}{Mipf parameter}

\item{...}{dots}
}
\value{
list
}
\description{
CellKeyFits
}
\examples{
my_km2 <- SSBtools::SSBtoolsData("my_km2")
CellKeyFits(my_km2, "freq", formula = ~(Sex + Age) * Municipality * Square1000m + Square250m)  
}
