% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellKeyFits.R
\name{CellKeyFits}
\alias{CellKeyFits}
\title{CellKeyFits}
\usage{
CellKeyFits(
  data,
  freqVar = NULL,
  rKeyVar = NULL,
  hierarchies = NULL,
  formula = NULL,
  dimVar = NULL,
  preAggregate = is.null(freqVar),
  xReturn = FALSE,
  extend0 = TRUE,
  limit = 1e-10,
  viaQR = FALSE,
  iter = 1000,
  eps = 0.01,
  tol = 1e-13,
  reduceBy0 = TRUE,
  reduceByColSums = TRUE,
  reduceByLeverage = FALSE,
  ...
)
}
\arguments{
\item{data}{data frame (inner cells)}

\item{freqVar}{Variable holding counts}

\item{rKeyVar}{Variable holding keys}

\item{hierarchies}{List of hierarchies}

\item{formula}{Model formula}

\item{dimVar}{Dimensional variables}

\item{preAggregate}{Aggregation}

\item{xReturn}{Dummy matrix in output when \code{TRUE}. To return crossTable as well, use \code{xReturn = 2}.}

\item{extend0}{Extend0}

\item{limit}{LSfitNonNeg parameter}

\item{viaQR}{LSfitNonNeg parameter}

\item{iter}{Mipf parameter}

\item{eps}{Mipf parameter}

\item{tol}{Mipf parameter}

\item{reduceBy0}{Mipf parameter}

\item{reduceByColSums}{Mipf parameter}

\item{reduceByLeverage}{Mipf parameter}

\item{...}{Further parameters to CellKey}
}
\value{
list
}
\description{
CellKeyFits
}
\examples{
z <- data.frame(geo  = c("Iceland", "Portugal", "Spain"), 
                eu = c("nonEU", "EU", "EU"),
                year = rep(c("2018","2019"), each = 3),
                freq = c(2,3,7,1,5,6), stringsAsFactors = FALSE)
z4 <- z[-c(1:2), ]

CellKeyFits(z4, "freq", formula = ~eu * year + geo, extend0 = FALSE)
CellKeyFits(z4, "freq", formula = ~eu * year + geo)
set.seed(123)
z4$keys <- runif(4)
CellKeyFits(z4, "freq", "keys", formula = ~eu * year + geo)                 

my_km2 <- SSBtools::SSBtoolsData("my_km2")

CellKeyFits(my_km2, "freq", formula = ~(Sex + Age) * Municipality * Square1000m + Square250m)  
}
